% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgrmhelperfunctions.R
\name{findMatrixElements}
\alias{findMatrixElements}
\title{Find row and column indices and matrix values for a vector of elements}
\usage{
findMatrixElements(pvElements, pvFlatVec)
}
\arguments{
\item{pvElements}{Vector of indices specifying elements to be searched for}

\item{pvFlatVec}{Flattened vector represenation of symmetric matrix}
}
\value{
lResultLoL     Result list of lists with components nRowIndex, nColIndex and nCoefficient
}
\description{
Given a vector of element indices of a flattened vector and the flattened
vector itself, find row and column indices and coefficients (matrix values)
for the specified elements.
}
\details{
Flattening a symmetric matrix corresponds to saving the elements of the lower
(or upper) triangular part of the matrix in a defined order in a vector. Given
such a flattened representation of a matrix one might search for the exact
location where in the original matrix certain elements did occur.
}

