% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgrmhelperfunctions.R
\name{findSingleMatElement}
\alias{findSingleMatElement}
\title{Find a single element of a symmetric matrix based on a flattened vector representation}
\usage{
findSingleMatElement(pnElementIdx, pvFlatVec)
}
\arguments{
\item{pnElementIdx}{Vector index of element to be searched for}

\item{pvFlatVec}{Flattened vector represenation of symmetric matrix}
}
\value{
lResultList    Result list with components nRowIndex, nColIndex and nCoefficient
}
\description{
Given a flattened vector representation of a symmatric matrix \code{findSingleMatElement}
computes the row and column indices in the original matrix of a given element. The element
is given by the index of the flattened vector. The flattened vector is a vector that contains
all elements of the lower triangular part of the matrix including all diagonal elements. The
order in which the elements are written to the vector is going through each row and copying
each element from column one up to the diagonal element of that row into the flattened
vector.
}

