% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgrmhelperfunctions.R
\name{flattenMatrixLowerTri}
\alias{flattenMatrixLowerTri}
\title{Flatten lower triangular part of a matrix into a vector}
\usage{
flattenMatrixLowerTri(pMatrix, psOrder = "rowwise", pbIncludeDiag = TRUE)
}
\arguments{
\item{pMatrix}{Matrix to be flattened}

\item{psOrder}{Order in which elements should be stored in the flattened vector}

\item{pbIncludeDiag}{whether or not to include diagonal elements of pMatrix}
}
\value{
vFlatVecResult   flattened vector
}
\description{
\code{flattenMatrixLowerTri} uses function \code{lower.tri()} and
\code{upper.tri()} to flatten the lower triangular part of a matrix
into a vector.
}

