% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgrmhelperfunctions.R
\name{nGetMatDim}
\alias{nGetMatDim}
\title{Compute matrix dimension based on the length of a flattened vector}
\usage{
nGetMatDim(pnLenFlatVec)
}
\arguments{
\item{pnLenFlatVec}{length of flattened vector}
}
\value{
nMatDimResult   dimension of symmetric matrix
}
\description{
The flattened vector contains all elements of the lower triangular
part of a symmetric matrix including all diagonal elements. Because
any row i of a quadratic matrix contains i elements below the
diagonal, including the diagonal element, the flattened vector of
an n * n symmetric matrix contains the sum of all natural numbers
up to and including n. That length has a closed form and corresponds
to n*(n+1)/2. Given the length of the flattened vector, we have
a quadratic equation for n, that can be solved. We are only interested
in the positive solution which is n = (-1 + \sqrt(1+8l))/2
}

