% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgrmreadfunctions.R
\name{ReadGRMBin}
\alias{ReadGRMBin}
\title{Reading Genetic Relationship Matrix from binary files}
\usage{
ReadGRMBin(prefix, AllN = F, size = 4)
}
\arguments{
\item{prefix}{common prefix of input files}

\item{AllN}{flag indicating whether to read 1 line (FALSE) or n(n+1)/2 lines from file prefix.N.bin}

\item{size}{wordsize that is passed to readBin}
}
\value{
list with diagonal elements, offdiagonals, ids and total number of individuals
}
\description{
The function \code{ReadGRMBin} was taken from
http://cnsgenomics.com/software/gcta/estimate_grm.html
and was only slightly modified, but in essence
the function here does the same thing. Given three
files which have names starting with the same prefix,
a list with the diagonal elements, the offdiagonal
elements and the ids is returned
}

