% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgrmreadfunctions.R
\name{readGRMBinToVector}
\alias{readGRMBinToVector}
\title{Read a GRM into a Vector}
\usage{
readGRMBinToVector(psInputPrefix, psInputPath = NULL, size = 4)
}
\arguments{
\item{psInputPrefix}{common prefix of input files}

\item{psInputPath}{path to input files}

\item{size}{word size passed to readBin}
}
\value{
vector with flattened represenation of GRM
}
\description{
Given two input files

\itemize{
\item a file with the flattened GRM in binary format
\item a file with indices
}

The data from the binary file are read using readLines and
are stored in a vector
}

