% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgrmhelperfunctions.R
\name{runCompleteTest}
\alias{runCompleteTest}
\title{Test all elements of flattened vector}
\usage{
runCompleteTest(pvFlatMatElem, pMatrix)
}
\arguments{
\item{pvFlatMatElem}{flattened vector}

\item{pMatrix}{original matrix}
}
\value{
dfResults       dataframe with comparison results
}
\description{
The test consists of searching for all elements in the flattened
vector and comparing the found elements with the original elements
in the matrix
}

